+++ __init__.py
@@ -1,1 +1,1 @@
-"""ZM-QQManager —— 功能全面的 QQ 群管理插件。"""
+"""ZM-QQGroupmgr —— 功能全面的 QQ 群管理插件。"""